Move `LiveDataExtensions.kt` to :app:ui:base
