+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/managefolders/FolderSettingsFragment.kt
@@ -10,8 +10,8 @@ import androidx.preference.Preference
 import com.fsck.k9.fragment.ConfirmationDialogFragment
 import com.fsck.k9.fragment.ConfirmationDialogFragment.ConfirmationDialogFragmentListener
 import com.fsck.k9.ui.R
+import com.fsck.k9.ui.base.livedata.observeNotNull
 import com.fsck.k9.ui.folders.FolderNameFormatter
-import com.fsck.k9.ui.observeNotNull
 import com.takisoft.preferencex.PreferenceFragmentCompat
 import org.koin.android.ext.android.inject
 import org.koin.androidx.viewmodel.ext.android.viewModel
