+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/settings/import/SettingsImportFragment.kt
@@ -16,7 +16,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import com.fsck.k9.activity.setup.OAuthFlowActivity
 import com.fsck.k9.ui.R
-import com.fsck.k9.ui.observeNotNull
+import com.fsck.k9.ui.base.livedata.observeNotNull
 import com.mikepenz.fastadapter.FastAdapter
 import com.mikepenz.fastadapter.adapters.ItemAdapter
 import org.koin.androidx.viewmodel.ext.android.activityViewModel
